refactor: redirect adapters imports to data layer
Redirect all quodeq.adapters.* imports to quodeq.data.* where
canonical implementations exist. The only remaining adapters import
is _file_lock which has no data/ equivalent yet.
