+++ tests/test_fs_dimensions_repository.py
@@ -1,6 +1,6 @@
 import json
 import pytest
-from quodeq.adapters.fs.dimensions_repository import FilesystemDimensionsRepository
+from quodeq.data.fs.dimensions_repository import FilesystemDimensionsRepository
 
 
 def test_fs_dimensions_repository_reads_dimension(tmp_path):
